Refactored teh CommunityPRsReportGenerator to separate data retrieval and report generation
